# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ticketing System API",
        "description": "API documentation for Ticketing System",
        "contact": {
            "email": "support@ticketingsystem.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/sales-report/export": {
            "post": {
                "tags": [
                    "Sales Report"
                ],
                "summary": "Export sales report data",
                "description": "Export sales report data.",
                "operationId": "70bedcacd87dd3250acc105933827b97",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "format": {
                                        "description": "Export format (csv, excel, pdf)",
                                        "type": "string",
                                        "example": "csv"
                                    },
                                    "search": {
                                        "description": "Search term",
                                        "type": "string",
                                        "example": "transaction"
                                    },
                                    "cashier": {
                                        "description": "Cashier ID filter",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "startDate": {
                                        "description": "Start date",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-01"
                                    },
                                    "endDate": {
                                        "description": "End date",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-12-31"
                                    },
                                    "promoter": {
                                        "description": "Promoter ID filter",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "rate": {
                                        "description": "Rate ID filter",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sales report exported successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sales-report/statistics": {
            "get": {
                "tags": [
                    "Sales Report"
                ],
                "summary": "Get sales report statistics",
                "description": "Get sales report statistics.",
                "operationId": "bfc513dbced098cdc57ec8689834d241",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "cashier",
                        "in": "query",
                        "description": "Cashier ID filter",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "startDate",
                        "in": "query",
                        "description": "Start date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "endDate",
                        "in": "query",
                        "description": "End date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "promoter",
                        "in": "query",
                        "description": "Promoter ID filter",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "rate",
                        "in": "query",
                        "description": "Rate ID filter",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sales statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/scan": {
            "post": {
                "tags": [
                    "Scan Management"
                ],
                "summary": "Store scan data from barcode scanner",
                "description": "Store scan data from barcode scanner.",
                "operationId": "e0781fbb9dcf44f5c10ebab712e04b96",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "vgdecoderesult": {
                                        "description": "Hex encoded scan result",
                                        "type": "string",
                                        "example": "313233343536"
                                    },
                                    "devicenumber": {
                                        "description": "Device number",
                                        "type": "string",
                                        "example": "001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Scan data stored successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Scan received"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid scan data"
                    }
                }
            }
        },
        "/api/scan/latest": {
            "get": {
                "tags": [
                    "Scan Management"
                ],
                "summary": "Get the latest scan data",
                "description": "Get the latest scan data.",
                "operationId": "f2038ed05bc0faa12cc751525d69c1b4",
                "responses": {
                    "200": {
                        "description": "Latest scan data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "description": "Decoded scan code",
                                            "type": "string",
                                            "example": "123456"
                                        },
                                        "device": {
                                            "description": "Device number",
                                            "type": "string",
                                            "example": "001"
                                        },
                                        "timestamp": {
                                            "description": "Scan timestamp",
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-01-01T12:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No scan data available"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}